+++ lib/flutter_rustore_billing.dart
@@ -4,9 +4,9 @@ import 'package:flutter_rustore_billing/pigeons/rustore.dart';
 class RustoreBillingClient {
   static final RustoreBilling _api = RustoreBilling();
 
-  static Future<String> initialize(String id, String prefix, bool debugLogs,
-      bool allowNativeErrorHadling) async {
-    return _api.initialize(id, prefix, debugLogs, allowNativeErrorHadling);
+  static Future<String> initialize(
+      String id, String prefix, bool debugLogs) async {
+    return _api.initialize(id, prefix, debugLogs);
   }
 
   static Future<bool> available() async {