+++ pigeons/rustore.dart
@@ -113,8 +113,7 @@ class DigitalShopGeneralError {
 @HostApi()
 abstract class RustoreBilling {
   @async
-  String initialize(
-      String id, String prefix, bool debugLogs, bool? allowNativeErrorHadling);
+  String initialize(String id, String prefix, bool debugLogs);
 
   @async
   bool available();